changes to exercises for weeks 1-3 and part of Assignment 1

+++ src/main/java/uk/cam/mrc/phm/homework/lectures/week3/DebuggingExercise.java
@@ -9,7 +9,7 @@ public class DebuggingExercise {
     }
 
     public static void printNumberSeries(int num){
-        for (byte i = 0; i <= num; i+=2){
+        for (int i = 0; i <= num; i+=2){
             System.out.println(i);
         }
     }
